pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fortran2rust"
version = "0.1.0"
description = "Pipeline to convert Fortran code to safe, idiomatic Rust"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0",
    "fparser>=0.1.4",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.7",
    "rich>=13.0",
    "questionary>=2.0",
    "jinja2>=3.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pytest"]

[project.scripts]
fortran2rust = "fortran2rust.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
